+++ pgactivity/queries/get_blocking_oldest.sql
@@ -32,10 +32,7 @@ SELECT
             pg_stat_activity.datname,
             pg_stat_activity.datid,
             pg_stat_activity.usename,
-            CASE WHEN pg_stat_activity.client_addr IS NULL
-                THEN 'local'
-                ELSE pg_stat_activity.client_addr::TEXT
-            END AS client,
+            pg_stat_activity.client_addr AS client,
             blocking.locktype,
             EXTRACT(epoch FROM (NOW() - pg_stat_activity.{duration_column})) AS duration,
             NULL AS state,
@@ -65,10 +62,7 @@ SELECT
             pg_stat_activity.datname,
             pg_stat_activity.datid,
             pg_stat_activity.usename,
-            CASE WHEN pg_stat_activity.client_addr IS NULL
-                THEN 'local'
-                ELSE pg_stat_activity.client_addr::TEXT
-            END AS client,
+            pg_stat_activity.client_addr AS client,
             blocking.locktype,
             EXTRACT(epoch FROM (NOW() - pg_stat_activity.{duration_column})) AS duration,
             NULL AS state,
@@ -98,10 +92,7 @@ SELECT
             pg_stat_activity.datname,
             pg_stat_activity.datid,
             pg_stat_activity.usename,
-            CASE WHEN pg_stat_activity.client_addr IS NULL
-                THEN 'local'
-                ELSE pg_stat_activity.client_addr::TEXT
-            END AS client,
+            pg_stat_activity.client_addr AS client,
             blocking.locktype,
             EXTRACT(epoch FROM (NOW() - pg_stat_activity.{duration_column})) AS duration,
             NULL AS state,