+++ pgactivity/queries/get_waiting_post_090200.sql
@@ -7,10 +7,7 @@ SELECT
       a.application_name AS application_name,
       a.datname AS database,
       a.usename AS user,
-      CASE WHEN a.client_addr IS NULL
-          THEN 'local'
-          ELSE a.client_addr::TEXT
-      END AS client,
+      a.client_addr AS client,
       pg_locks.mode AS mode,
       pg_locks.locktype AS type,
       pg_locks.relation::regclass AS relation,